Fixed linter. Cleanup.
